v0.1.2-beta [Vectors]
New way to see the library, using the vector library to improve its
efficiency and versatility considerably.
Now we have a lot of work ahead to adapt and redesign all the code
necessary.
I will review all methods one by one to improve with the help of the
library Vector.
Remember that if you do not want to use a library that makes imports
from other libraries, you have the version 0.1.1 available to use
(without any detected bug)